requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.2.0
schedule==1.2.0
click==8.1.0
pyyaml==6.0.1